quick fix for response issues
In the statistics object, 'info': ctf_key, is added to the object. For modifications with multiple columns, this works correctly and adds like ['interest', 'principal']. However, for modifications with just one column (i.e. interest) it sends a string instead of an array with length one.
Second, the order of columns in <model>.columns.<modification label> doesn't have the modified column next to the original.